test(scripts): declare the shared real-Mongo budget in the migration suites
The CI "misc" leg fails intermittently on the real-DB migration suites in
this package - a 30s timeout that asserts nothing and goes green on re-run
of the same commit.

Booting mongod costs a cold start that scales with runner contention
(Mongoose builds every model's index set on connect; autoIndex cannot be
disabled). MONGO_TEST_TIMEOUT_MS documents the measured spread as 3-12s
healthy and 16-22s under shard parallelism; these three suites take
14.5-18.6s locally. The shard's budget is 30s for tests and 30s for hooks,
both sized for unit tests, so the suites sat inside that spread and crossed
it on scheduling luck.

apps/client already solved this: declare the shared 60s budget per file and
guard it structurally. This adopts the same pattern here rather than raising
the shard default again (15s to 30s was the last attempt) - a shard-wide
raise would hand the same slack to ~900 unit tests that must stay tight.

Extract the client guard's AST auditor so one implementation serves both
shards, and strip the 24 literal timeout arguments that pinned individual
tests and hooks back to the old budget.
